Move the caching logic to the handler in pkg/power/handler  - This allows all URL's to be cached if more than 1 URL is specified  - Alter the cache_file_name to cache_file_prefix. The URL index is added to the    prefix to produce the actual cache_file_name. The default is cached_forecast, resulting    in a cache file of cached_fiecast.0 for the first cache file.
