+++ pkg/cmd/estimate.go
@@ -14,7 +14,7 @@ import (
 var e_url string
 var e_apiKey string
 var e_cache_forecast bool
-var e_cache_file_name string
+var e_cache_file_prefix string
 var e_cache_time int32
 
 var estCmd = &cobra.Command{
@@ -26,7 +26,7 @@ var estCmd = &cobra.Command{
 		e_apiKey = viper.GetString("apikey")
 		fronius_ip = viper.GetString("fronius_ip")
                 e_cache_forecast = viper.GetBool("cache_forecast")
-                e_cache_file_name = viper.GetString("cache_file_name")
+                e_cache_file_prefix = viper.GetString("cache_file_prefix")
                 e_cache_time = viper.GetInt32("cache_time")
 
 		err := CheckEstimate(e_apiKey, e_url, fronius_ip)
@@ -43,16 +43,16 @@ func init() {
 	estCmd.Flags().StringVarP(&e_url, "url", "u", "", "Set the forecast URL. For multiple URLs, use a comma (,) to separate them")
 	estCmd.Flags().StringVarP(&e_apiKey, "apikey", "k", "", "set APIKEY")
 	estCmd.Flags().StringVarP(&fronius_ip, "fronius_ip", "H", "", "set FRONIUS_IP")
-        estCmd.Flags().BoolVarP(&e_cache_forecast, "cache_forecast", "n", false, "CACHE_FORECAST (default false)")
-        estCmd.Flags().StringVarP(&e_cache_file_name, "cache_file_name", "f", "cached_forecast.json", "CACHE_FILE_NAME (default 'cached_forecast.json')")
-        estCmd.Flags().Int32VarP(&e_cache_time, "cache_time", "l", 7200, "CACHE_TIME (default 7200)")
+  estCmd.Flags().BoolVarP(&e_cache_forecast, "cache_forecast", "n", false, "CACHE_FORECAST (default false)")
+  estCmd.Flags().StringVarP(&e_cache_file_prefix, "cache_file_prefix", "f", "cached_forecast", "CACHE_FILE_NAME (default 'cached_forecast')")
+  estCmd.Flags().Int32VarP(&e_cache_time, "cache_time", "l", 7200, "CACHE_TIME (default 7200)")
 
 	viper.BindPFlag("url", estCmd.Flags().Lookup("url"))
 	viper.BindPFlag("apikey", estCmd.Flags().Lookup("apikey"))
 	viper.BindPFlag("fronius_ip", estCmd.Flags().Lookup("fronius_ip"))
-        viper.BindPFlag("cache_forecast", estCmd.Flags().Lookup("cache_forecast"))
-        viper.BindPFlag("cache_file_name", estCmd.Flags().Lookup("cache_file_name"))
-        viper.BindPFlag("cache_time", estCmd.Flags().Lookup("cache_time"))
+  viper.BindPFlag("cache_forecast", estCmd.Flags().Lookup("cache_forecast"))
+  viper.BindPFlag("cache_file_prefix", estCmd.Flags().Lookup("cache_file_prefix"))
+  viper.BindPFlag("cache_time", estCmd.Flags().Lookup("cache_time"))
 
 	rootCmd.AddCommand(estCmd)
 }
